function_source/requirements.txt
functions-framework==3.*
google-cloud-vision==3.*
google-cloud-storage==2.*
google-cloud-aiplatform>=1.38
Pillow==10.*
python-dotenv
requests
